+++ ujin/adapt/__init__.py
@@ -2,6 +2,7 @@
 from ujin.adapt.backoff import Backoff, CircuitBreaker
 from ujin.adapt.concurrency import AIMDLimiter, TokenBucket
 from ujin.adapt.interval import AdaptiveInterval
+from ujin.adapt.signals import PolicySignals, SignalAdvisor, derive_signals
 from ujin.adapt.site_store import HostRecord, SiteStore
 from ujin.adapt import jitter
 
@@ -14,4 +15,7 @@ __all__ = [
     "jitter",
     "SiteStore",
     "HostRecord",
+    "PolicySignals",
+    "derive_signals",
+    "SignalAdvisor",
 ]